refactor: persistência - documentos, repositórios e mappers atualizados
- Documentos com @TypeAlias para persistência polimórfica
- UserDocument com Map<academiaId, Set<roleId>> para roles
- ClienteDocument com índices compostos
- AvaliacaoFisicaDocument com discrimador _class
- Mappers com conversão bidirecional Document ↔ Domain
- TesteVo2MaxMapperRegistry para mapeamento de testes variados
- Removido AvaliadorDocument, AvaliadorMapper, AvaliadorRepository
- Removido PermissionDocument e PermissionRepository

+++ src/main/java/com/prosup/proinsight/infrastructure/persistence/document/MedicaoImcDocument.java
@@ -8,7 +8,7 @@ public class MedicaoImcDocument extends MedicaoDocument {
 
     private Integer massaCorporalGramas;
     private Integer alturaCm;
-    private Integer imcCalculado;
+    private Double imcCalculado;
     private String classificacaoImc;
 
     public MedicaoImcDocument() {
@@ -31,11 +31,11 @@ public class MedicaoImcDocument extends MedicaoDocument {
         this.alturaCm = alturaCm;
     }
 
-    public Integer getImcCalculado() {
+    public Double getImcCalculado() {
         return imcCalculado;
     }
 
-    public void setImcCalculado(Integer imcCalculado) {
+    public void setImcCalculado(Double imcCalculado) {
         this.imcCalculado = imcCalculado;
     }
